Validate signature minimum feature level
This also restructures the Validate trait to decouple it from
Signature, and defines specific validation steps in the Signature
trait that can be optionally implemented or overridden by individual
Signature implementations.

The `feature_levels` method was also renamed to `computed_feature_level`
for clarity, and now returns an optional Range rather than a tuple
(for which the maximum was always None).