sig-validator: Added support for Y-type signatures in WDB files.
Y-type signatures allow a regular expression to be defined for a real
URI that should always be ignored for phishing checks, regardless of
what the displayed URI is. An example of this would be a safe browsing
service that will always rewrite URIs to point to the service to verify
that a URI is safe before redirecting the client to the real URI. In
this case, a regex to cover the safe browsing URI would be added as a
Y-type signature.

CLAM-2425